remove TempFile and stub out some docs
